+++ src/visitors/static/add-static-proxy.js
@@ -1,4 +1,4 @@
-let {toLogicalID} = require('@architect/utils')
+let toLogicalID = require('../../to-logical-id')
 
 module.exports = function addStatic(arc, template) {