+++ src/visitors/static/index.js
@@ -1,6 +1,4 @@
-//let {toLogicalID} = require('@architect/utils')
 let addStatic = require('./add-static-proxy')
-let addMocks = require('./add-static-mocks')
 
 /**
  * visit arc.static and merge in AWS::Serverless resources
@@ -43,7 +41,6 @@ module.exports = function statics(arc, template) {
   // if an api is defined then add _static proxy and attempt to serialize ./public
   if (arc.http) {
     template = addStatic(arc, template)
-    template = addMocks(arc, template)
   }
 
   return template